Move envs to their own files
