+++ dhall/src/semantics/mod.rs
@@ -3,4 +3,5 @@ pub mod nze;
 pub mod phase;
 pub mod tck;
 pub(crate) use self::core::*;
+pub(crate) use self::nze::*;
 pub(crate) use self::tck::*;
