+++ dhall/src/semantics/nze/mod.rs
@@ -1,2 +1,2 @@
-pub mod nzexpr;
-pub(crate) use nzexpr::*;
+pub mod env;
+pub(crate) use env::*;
