+++ dhall/src/semantics/tck/mod.rs
@@ -1,4 +1,6 @@
 pub mod context;
+pub mod env;
 pub mod tyexpr;
 pub mod typecheck;
+pub(crate) use env::*;
 pub(crate) use tyexpr::*;
